+++ app/src/androidTest/java/com/github/sdp/mediato/GreetingActivityTest.java
@@ -6,25 +6,25 @@ import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 
 import android.content.Intent;
-
 import androidx.test.core.app.ActivityScenario;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
-
+import com.github.sdp.mediato.ui.GreetingActivity;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 @RunWith(AndroidJUnit4.class)
 public class GreetingActivityTest {
 
-    @Test
-    public void CheckGreetingMessage(){
-        Intent myIntent = new Intent(ApplicationProvider.getApplicationContext(), GreetingActivity.class);
-        myIntent.putExtra("mainName","Michel");
+  @Test
+  public void CheckGreetingMessage() {
+    Intent myIntent = new Intent(ApplicationProvider.getApplicationContext(),
+        GreetingActivity.class);
+    myIntent.putExtra("mainName", "Michel");
 
-        try(ActivityScenario test = ActivityScenario.launch(myIntent)) {
-            onView(withId(R.id.greetingMessage)).check(matches(withText("Hello Michel!")));
-        }
+    try (ActivityScenario test = ActivityScenario.launch(myIntent)) {
+      onView(withId(R.id.greetingMessage)).check(matches(withText("Hello Michel!")));
     }
+  }
 
 }
