+++ app/src/main/AndroidManifest.xml
@@ -1,38 +1,38 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+  xmlns:tools="http://schemas.android.com/tools">
 
-    <application
-        android:allowBackup="true"
-        android:dataExtractionRules="@xml/data_extraction_rules"
-        android:fullBackupContent="@xml/backup_rules"
-        android:icon="@mipmap/ic_launcher"
-        android:label="@string/mt_app_name"
-        android:supportsRtl="true"
-        android:theme="@style/Theme.MediaTo"
-        android:windowSoftInputMode="adjustResize"
-        tools:targetApi="31">
-	<activity
-            android:name=".NewItemActivity"
-            android:exported="false" />
-        <activity
-            android:name=".TestingActivity"
-            android:exported="false" />
-        <activity
-            android:name=".GreetingActivity"
-            android:exported="false" />
-        <activity
-            android:name=".MainActivity"
-            android:exported="false" />
+  <application
+    android:allowBackup="true"
+    android:dataExtractionRules="@xml/data_extraction_rules"
+    android:fullBackupContent="@xml/backup_rules"
+    android:icon="@mipmap/ic_launcher"
+    android:label="@string/mt_app_name"
+    android:supportsRtl="true"
+    android:theme="@style/Theme.MediaTo"
+    android:windowSoftInputMode="adjustResize"
+    tools:targetApi="31">
+    <activity
+      android:name=".ui.NewItemActivity"
+      android:exported="false" />
+    <activity
+      android:name=".ui.TestingActivity"
+      android:exported="false" />
+    <activity
+      android:name=".ui.GreetingActivity"
+      android:exported="false" />
+    <activity
+      android:name=".ui.MainActivity"
+      android:exported="false" />
 
-        <activity
-            android:name=".NewProfileActivity"
-            android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
+    <activity
+      android:name=".ui.NewProfileActivity"
+      android:exported="true">
+      <intent-filter>
+        <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-    </application>
+        <category android:name="android.intent.category.LAUNCHER" />
+      </intent-filter>
+    </activity>
+  </application>
 </manifest>
