+++ app/src/main/java/com/github/sdp/mediato/ui/CreateProfileFragment.java
@@ -1,4 +1,4 @@
-package com.github.sdp.mediato;
+package com.github.sdp.mediato.ui;
 
 import android.content.Intent;
 import android.net.Uri;
@@ -13,6 +13,7 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import com.github.javafaker.Faker;
+import com.github.sdp.mediato.R;
 import com.github.sdp.mediato.data.Database;
 import com.github.sdp.mediato.formats.Dates;
 import com.github.sdp.mediato.model.Location;
