+++ app/src/main/java/com/github/sdp/mediato/ui/MainActivity.java
@@ -1,4 +1,4 @@
-package com.github.sdp.mediato;
+package com.github.sdp.mediato.ui;
 
 import android.os.Bundle;
 import android.widget.Toast;
@@ -6,7 +6,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentTransaction;
-import com.github.sdp.mediato.databinding.ActivityMainBinding;
+import com.github.sdp.mediato.R;
 
 /**
  * The main activity of the app that displays a bottom navigation bar and manages the navigation
