+++ app/src/main/java/com/github/sdp/mediato/ui/SearchFragment.java
@@ -1,4 +1,4 @@
-package com.github.sdp.mediato;
+package com.github.sdp.mediato.ui;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -7,6 +7,7 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 import androidx.fragment.app.Fragment;
+import com.github.sdp.mediato.R;
 import java.util.ArrayList;
 
 public class SearchFragment extends Fragment {
