+++ app/src/main/res/layout/fragment_search.xml
@@ -4,7 +4,7 @@
   android:layout_width="match_parent"
   android:layout_height="match_parent"
   android:background="@color/colorPrimary"
-  tools:context=".SearchFragment">
+  tools:context=".ui.SearchFragment">
 
   <!-- TODO: Update blank fragment layout -->
   <TextView
